A few changes towards error handling. Not yet completed.

+++ message/src/main/java/org/ietf/nea/pb/batch/PbBatchHeaderBuilder.java
@@ -6,7 +6,7 @@ import de.hsbremen.tc.tnc.tnccs.batch.TnccsBatchHeaderBuilder;
 
 public interface PbBatchHeaderBuilder extends TnccsBatchHeaderBuilder {
 
-	public abstract PbBatchHeaderBuilder setVersion(byte version)
+	public abstract PbBatchHeaderBuilder setVersion(short version)
 			throws RuleException;
 
 	public abstract PbBatchHeaderBuilder setDirection(byte direction)
