+++ message/src/main/java/org/ietf/nea/pb/validate/rules/BatchVersion.java
@@ -6,7 +6,7 @@ import org.ietf.nea.pb.validate.enums.PbErrorCauseEnum;
 
 public class BatchVersion {
 
-	public static void check(final byte actuallVersion, final byte supportedVersion) throws RuleException{
+	public static void check(final short actuallVersion, final short supportedVersion) throws RuleException{
         if(actuallVersion != supportedVersion){
             throw new RuleException("The version "+actuallVersion+" is not supported.",true,PbMessageErrorCodeEnum.IETF_UNSUPPORTED_VERSION.code(),PbErrorCauseEnum.BATCH_VERSION_NOT_SUPPORTED.number(),actuallVersion);
         }
